I now understand how alloc user counter works. Include userid in WHERE clause when deleting inbox notifications.
git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@8551 1dfba125-df10-0410-8a77-af1a8d6d074f

+++ cgi-bin/LJ/NotificationInbox.pm
@@ -134,7 +134,7 @@ sub delete_from_queue {
     my $u = $self->u
         or die "No user object";
 
-    $u->do("DELETE FROM notifyqueue WHERE qid=?", undef, $qid);
+    $u->do("DELETE FROM notifyqueue WHERE userid=? AND qid=?", undef, $u->id, $qid);
     die $u->errstr if $u->err;
 
     # invalidate caches
